requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
packaging==24.0
python-magic==0.4.27
Werkzeug==3.0.2
zipp==3.18.1
requests~=2.31.0
python-dotenv==1.0.1
PyJWT~=2.8.0
apscheduler==3.10.4
pybase64==1.3.2
flask_sqlalchemy==3.1.1
mysqlclient==2.2.4
watchdog==4.0.0
libmagic==1.0
mimetype==0.1.5
bcrypt~=3.2.0
mysql==0.0.3
mysql-connector-python~=8.3.0
pytz~=2023.3.post1
sqlalchemy~=2.0.30